refactoring abstract classes into package
