+++ src/test/java/model/ConfigurationGeneralPage.java
@@ -1,11 +1,12 @@
 package model;
 
+import model.base.BasePage;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
 import org.openqa.selenium.support.ui.ExpectedConditions;
 
-public class ConfigurationGeneralPage extends BasePage{
+public class ConfigurationGeneralPage extends BasePage {
     @FindBy(xpath = "//button[@type='submit']")
     private WebElement saveButton;
