+++ src/test/java/model/ConfigureUserPage.java
@@ -1,10 +1,11 @@
 package model;
 
+import model.base.BasePage;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
 
-public class ConfigureUserPage extends BasePage{
+public class ConfigureUserPage extends BasePage {
 
     @FindBy(css = "input[name='_.fullName']")
     private WebElement inputFieldFullUserName;
