+++ src/test/java/model/CreateUserPage.java
@@ -1,10 +1,11 @@
 package model;
 
+import model.base.BasePage;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
 
-public class CreateUserPage extends BasePage{
+public class CreateUserPage extends BasePage {
 
     @FindBy(id = "username")
     private WebElement username;
