+++ src/test/java/model/DeleteMultibranchPipelinePage.java
@@ -1,10 +1,11 @@
 package model;
 
+import model.base.BasePage;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
 
-public class DeleteMultibranchPipelinePage extends BasePage{
+public class DeleteMultibranchPipelinePage extends BasePage {
     @FindBy(xpath = "//button[@type = 'submit']")
     private WebElement submitButton;
