+++ src/test/java/model/DeleteViewConfirmationPage.java
@@ -1,10 +1,11 @@
 package model;
 
+import model.base.MyViewSideMenuFrame;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
 
-public class DeleteViewConfirmationPage extends MyViewSideMenuFrame{
+public class DeleteViewConfirmationPage extends MyViewSideMenuFrame {
 
     @FindBy(id = "yui-gen1-button")
     private WebElement yesButtonDeleteView;
