+++ src/test/java/model/GlobalToolConfigurationPage.java
@@ -1,6 +1,6 @@
 package model;
 
-import org.openqa.selenium.By;
+import model.base.BasePage;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
