+++ src/test/java/model/HomePage.java
@@ -1,5 +1,6 @@
 package model;
 
+import model.base.SideMenuFrame;
 import org.openqa.selenium.By;
 import org.openqa.selenium.Keys;
 import org.openqa.selenium.WebDriver;
