+++ src/test/java/model/MultiConfigurationProjectStatusPage.java
@@ -1,12 +1,13 @@
 package model;
 
+import model.base.BasePage;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
 import org.openqa.selenium.support.ui.ExpectedConditions;
 
-public class MultiConfigurationProjectStatusPage extends BasePage{
+public class MultiConfigurationProjectStatusPage extends BasePage {
 
     @FindBy(css = "#breadcrumbs li a")
     private WebElement dashboard;
