+++ src/test/java/model/MyViewsPage.java
@@ -1,5 +1,6 @@
 package model;
 
+import model.base.MyViewSideMenuFrame;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
