+++ src/test/java/model/NewViewPage.java
@@ -1,5 +1,6 @@
 package model;
 
+import model.base.MyViewSideMenuFrame;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
