+++ src/test/java/model/PeoplePage.java
@@ -1,14 +1,14 @@
 package model;
 
+import model.base.BasePage;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
 
 import java.util.ArrayList;
-import java.util.Collection;
 import java.util.List;
 
-public class PeoplePage extends BasePage{
+public class PeoplePage extends BasePage {
 
     @FindBy(xpath = "//tbody/tr/td")
     private List<WebElement> usersListInPeople;
