+++ src/test/java/model/RenameItemPage.java
@@ -1,10 +1,11 @@
 package model;
 
+import model.base.BasePage;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
 
-public class RenameItemPage extends BasePage{
+public class RenameItemPage extends BasePage {
 
     public RenameItemPage(WebDriver driver) {
         super(driver);
