+++ src/test/java/model/ViewPage.java
@@ -1,6 +1,6 @@
 package model;
 
-import org.openqa.selenium.By;
+import model.base.MyViewSideMenuFrame;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
@@ -8,7 +8,7 @@ import org.openqa.selenium.support.FindBy;
 import java.util.List;
 import java.util.stream.Collectors;
 
-public class ViewPage extends MyViewSideMenuFrame{
+public class ViewPage extends MyViewSideMenuFrame {
 
     @FindBy(css = "tr td a.model-link")
     private List<WebElement> jobList;
