+++ src/test/java/model/base/BasePage.java
@@ -1,4 +1,4 @@
-package model;
+package model.base;
 
 import org.openqa.selenium.WebDriver;
 import runner.BaseModel;
