+++ src/test/java/model/base/MainPage.java
@@ -1,5 +1,8 @@
-package model;
+package model.base;
 
+import model.HomePage;
+import model.MyViewsPage;
+import model.base.BasePage;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
